Move file-icon to file-explorer/views
