+++ src/components/tab-bar/tab-drag-preview.tsx
@@ -1,6 +1,6 @@
 import { Database, Package, Pin } from "lucide-react";
+import FileIcon from "../../file-explorer/views/file.icon";
 import type { Buffer } from "../../types/buffer";
-import FileIcon from "../file-icon";
 
 interface TabDragPreviewProps {
   x: number;
