+++ src/file-explorer/views/file.icon.tsx
